+++ src/shared/ui/Select/Option.tsx
@@ -1,5 +1,5 @@
 'use client';
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import { useSelectContext } from './Select.context';
